feat:basic form

+++ src/Controller/ServiceController.php
@@ -4,10 +4,14 @@ namespace App\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 use App\Repository\AuthorRepository;
 use App\Repository\BookRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use App\Form\AuthorType;
+use App\Entity\Author;
+use Doctrine\Persistence\ManagerRegistry;
 
 #[Route('/author')]
 class ServiceController extends AbstractController
@@ -40,26 +44,43 @@ class ServiceController extends AbstractController
             'authors' => $authors,
         ]);
     }
-    
-    #[Route('/display2', name:"app_display2")]
-    public function display2(BookRepository $rep): Response
-    {
-
-        $books = $rep->findAll();
-        return $this->render('book/display.html.twig', [
-            'books' => $books,
-        ]);
-    }
-
 
     #[Route('/delete/{id}', name:"delete")]
     public function delete(AuthorRepository $rep , $id, EntityManagerInterface $em): Response
     {
         $author = $rep->find($id);
 
+
+    if (!$author) {
+        $this->addFlash('error', 'Author not found');
+        return $this->redirectToRoute("app_display");
+    }
+
         $em->remove($author);
         $em->flush();
         return $this->redirectToRoute("app_display");
+    }
+
+    #[Route('/add', name:"add_author")]
+    public function addAuthor(ManagerRegistry $doctrine, Request $request): Response
+    {
+        $em = $doctrine->getManager();
+        $author = new author(); 
+        $form = $this->createForm(AuthorType::class, $author);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em->persist($author);
+            $em->flush();
+            
+        return $this->redirectToRoute("app_display");
+        
+        }
+
+        return $this->render('author/add.html.twig', [
+            'f' => $form
+        ]);
+
+
     }
 
 }
